src/requirements.txt
# BizFlow - Digital transformation platform for small businesses
# Flask Clean Architecture

# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Database
PyMySQL==1.1.0
SQLAlchemy>=2.0.0

# Redis (optional for realtime)
redis==5.0.1

# Validation & Serialization
marshmallow==3.21.0

# Security
bcrypt==4.1.2
python-dotenv==1.0.0

# API Documentation
flasgger==0.9.7.1
flask-swagger-ui==4.11.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Production server
gunicorn==21.2.0